+++ publish/CHANGELOG.md
@@ -1,6 +1,13 @@
+- v1.51
+  - Adds a new setting "Use blueprint chance" to use the object chance from blueprints.
+  - Adds a new setting "Save simpler blueprints" to only save the most important data.
+  - Changes the "Ignored ids" setting to also prevent the connecting with the `connect` parameter.
+  - Fixes compatibility issue with Building mod.
+  - Removes description from saved blueprints (was bugged and it's not really needed).
+
 - v1.50
   - Adds a new value "Legacy" to the setting "Set invulnerability".
-  - Fixes repairing not working with the setting "Set invulnerability".
+  - Fixes repairing not working with the setting "Set invulnerability" (some cases, not all).
   - Fixes changing build tab with Q/E not selecting the new object.
   - Fixes build menu showing old selection when selecting existing objects.
   - Fixes copying not always copying the rotation (for example active furnaces).